feat(function-calling-abstraction): move function calling utils to the `FunctionCalling` folder.
