+++ dotnet/src/Connectors/Connectors.AzureOpenAI/Core/AzureClientCore.cs
@@ -9,7 +9,7 @@ using Azure.Core;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
 #pragma warning disable IDE0005 // Using directive is unnecessary
-using Microsoft.SemanticKernel.Connectors.AI;
+using Microsoft.SemanticKernel.Connectors.FunctionCalling;
 #pragma warning restore IDE0005 // Using directive is unnecessary
 using Microsoft.SemanticKernel.Connectors.OpenAI;
 using Microsoft.SemanticKernel.Http;
